Add Support for PageRank Seed Vectors (#3)
Summary:
the cred calculation is defined by a Markov Mixing process, by introducing the seed vector and teleportation parameter alpha, the Markov mixing process is augmented with a source of cred originating from the seed vector. The resulting algorithm is the generalized variation of Pagerank, allowing computation of both canonical PageRank where the seed vector is the uniform distribution and personalized PageRank where the seed vector is an indicator distribution. It is still possible to get the simple markov chain solution by setting alpha = 0.

Note that this changes the Markov process state update, but does not provide updates to the APIs. All existing behavior is unchanged because alpha is always set to 0.


Test Plan:

Existing tests have been extended to include passing alpha = 0 to reproduce exisiting test cases for the simple Markov Process. Addition test cases include
 - Verifying that resulting stationary distribution is unaffected by seed when alpha = 0
 - Verifying that resulting stationary distribution is precisely equal to seed when alpha = 1
 - Verifying that the resulting stationary distribution is linear in the seed vector
 - Verifying that the correct stationary distribution is computed for non-zero alpha 
 - verify that the algorithm converges immediately when the initialDistribution is the stationary distribution
 - verify that the changing the initialDistribution does not change the stationary distribution