PagerankGraph allows specifying a seed vector
This commit modifies PagerankGraph so the user can now specify a seed
vector (along with an alpha) to modify PageRank's behavior. For example,
you could use it to start with a particular node as seed, and then
explore which parts of the graph have cred relative to that starting
node.

There are four kinds of seed strategies laid out:
- "NO_SEED", in which case alpha is 0 and no seeding occurs
- "UNIFORM_SEED", allows manually specifying alpha but always seeds back
  to a uniform vector
- "SELECTED_SEED", allows selecting a set of nodes to be the seed
  vector, the initial weight is split evenly between them.
- "SPECIFIED_SEED", which allows explicitly setting weights for each
  node, which will then be re-normalized

In principle, we could avoid all these types by just passing in alpha
and a node-weight map, with the following equivalencies:
- If alpha=0, then it's equivalent to NO_SEED
- If alpha=k and the map is empty, then it's equivalent to UNIFORM_SEED
- If alpha=k and the map has a number of nodes all with weight 1, then
  it's equivalent to SELECTED_SEED
- If alpha=k and the map has a number of nodes with arbitrary
  nonnegative weights, then it's equivalent to SPECIFIED_SEED

I think this API would be cleaner (and a lot less redundant) so I
anticipate that we will change this before merging it into
sourcecred/sourcecred master.