Finish using PagerankParams (#6)
We were a little sloppy in sourcecred/odyssey-hackathon#5, and didn't
actually use the new option argument in the primary exported API from
markovChain. This commit fixes that oversight. As expected, we had to
refactor all the Pagerank consumers now that we changed the primary API.

Test plan: `yarn test`

Paired with @mzargham